+++ src/lib/bindings.ts
@@ -353,9 +353,9 @@ async getPlayerNameCommand(uid: number) : Promise<Result<string | null, string>>
     else return { status: "error", error: e  as any };
 }
 },
-async savePacketCaptureSettings(method: string, npcapDevice: string) : Promise<Result<null, string>> {
+async savePacketCaptureSettings(npcapDevice: string) : Promise<Result<null, string>> {
     try {
-    return { status: "ok", data: await TAURI_INVOKE("save_packet_capture_settings", { method, npcapDevice }) };
+    return { status: "ok", data: await TAURI_INVOKE("save_packet_capture_settings", { npcapDevice }) };
 } catch (e) {
     if(e instanceof Error) throw e;
     else return { status: "error", error: e  as any };
@@ -929,4 +929,3 @@ import {
 export type Result<T, E> =
 	| { status: "ok"; data: T }
 	| { status: "error"; error: E };
-